requirements.txt
asgiref==3.8.1
attrs==25.3.0
Authlib==1.3.1
autobahn==24.4.2
Automat==25.4.16
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.1
dj-database-url==2.1.0
Django==5.2.3
django-cors-headers==4.4.0
django-oauth-toolkit==3.0.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
filelock==3.19.1
fsspec==2025.7.0
huggingface-hub==0.34.4
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
jwcrypto==1.5.6
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==6.2.0
regex==2025.8.29
requests==2.31.0
safetensors==0.6.2
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.6.0
zope.interface==7.2
